pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jhosua Lascarro", email = "168266147+Jhosua-Lascarro@users.noreply.github.com" },
]
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
]

[project.scripts]
app = "app:main"

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = ["ruff>=0.15.1", "ty>=0.0.17"]
